requirements.txt
alembic==1.13.3
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.2.0
billiard==4.2.1
black==24.10.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.2
fastapi-cache2==0.2.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
kombu==5.4.2
Mako==1.3.5
MarkupSafe==3.0.1
mypy-extensions==1.0.0
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
pendulum==3.0.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
redis==5.2.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.40.0
time-machine==2.16.0
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.0
vine==5.1.0
wcwidth==0.2.13
